+++ package.json
@@ -4,18 +4,17 @@
   "dependencies": {
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
+    "@reduxjs/toolkit": "^1.9.7",
     "@swc/plugin-emotion": "^2.5.92",
     "@types/react": "^18.2.31",
     "@types/react-dom": "^18.2.14",
     "axios": "^1.5.1",
-    "lottie-web": "^5.12.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hot-toast": "^2.4.1",
     "react-redux": "^8.1.3",
     "react-router-dom": "^6.17.0",
     "redux": "^4.2.1",
-    "redux-thunk": "^2.4.2",
     "typescript": "^5.2.2",
     "vite": "^4.5.0",
     "vite-plugin-svgr": "^4.1.0",
