+++ src/components/CheckBox/styled.ts
