+++ src/components/DownLoadCountSlider/styled.ts
