+++ src/components/FileListBox/styled.ts
