+++ src/components/FolderListBox/styled.ts
