+++ src/components/Progress/styled.ts
