+++ src/components/RangeSlider/index.ts
