--- src/components/common/index.ts
@@ -1,12 +1,0 @@
-export * from './CheckBox';
-export * from './PasswordInput';
-export * from './UpLoadButton';
-export * from './FileFind';
-export * from './DownLoadCountSlider';
-export * from './Navbar';
-export * from './Button';
-export * from './FileListBox';
-export * from './FolderListBox';
-export * from './Progress';
-export * from './SkeletonUIBox';
-export * from './ExpireTime';
