+++ src/index.tsx
@@ -1,17 +1,13 @@
-import React from 'react';
 import { Global } from '@emotion/react';
-import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
-
-import { App } from './App';
 import { globalStyle } from './styles';
+import { createRoot } from 'react-dom/client';
+import App from './App';
 
-const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
+const root = createRoot(document.getElementById('root') as HTMLElement);
 root.render(
-  <React.StrictMode>
-    <BrowserRouter>
-      <Global styles={globalStyle} />
-      <App />
-    </BrowserRouter>
-  </React.StrictMode>
+  <BrowserRouter>
+    <Global styles={globalStyle} />
+    <App />
+  </BrowserRouter>
 );
