+++ src/state/store.ts
@@ -1,6 +1,6 @@
-import { applyMiddleware, createStore } from 'redux';
-import thunk from 'redux-thunk';
+import { configureStore } from '@reduxjs/toolkit';
+import { rootReducer } from './reducers';
 
-import reducers from './reducers';
-
-export const store = createStore(reducers, {}, applyMiddleware(thunk));
+export const store = configureStore({
+  reducer: rootReducer,
+});
